Moved all the files of AudioUseCase package to audio_use_case package. Deleted AudioUseCasePackage.
