+++ src/main/java/audio_use_case/AudioBytesFactory.java
@@ -1,4 +1,4 @@
-package AudioUseCase;
+package audio_use_case;
 
 import com.google.protobuf.ByteString;
