+++ src/main/java/audio_use_case/AudioConverter.java
@@ -1,13 +1,8 @@
-package AudioUseCase;
+package audio_use_case;
 
-import com.google.api.gax.core.FixedCredentialsProvider;
-import com.google.auth.oauth2.ServiceAccountCredentials;
 import com.google.cloud.speech.v1.*;
-import com.google.cloud.speech.v1.RecognitionConfig.AudioEncoding;
 import com.google.protobuf.ByteString;
-import org.apache.commons.codec.language.bm.Lang;
 
-import java.io.FileInputStream;
 import java.io.IOException;
 import java.util.List;
